Move customHook of useFilter to respective file
